feat: consolidate comp structs
