+++ crates/cdk/src/wallet/saga/mod.rs
@@ -18,8 +18,11 @@ use std::collections::VecDeque;
 use std::sync::Arc;
 
 use async_trait::async_trait;
+use cdk_common::database::{self, WalletDatabase};
 use tokio::sync::Mutex;
+use tracing::instrument;
 
+use crate::nuts::{PublicKey, State};
 use crate::Error;
 
 /// Trait for compensating actions in the saga pattern.
@@ -83,6 +86,94 @@ pub async fn clear_compensations(compensations: &Compensations) {
 /// Add a compensating action to the front of the queue (LIFO order).
 pub async fn add_compensation(compensations: &Compensations, action: Box<dyn CompensatingAction>) {
     compensations.lock().await.push_front(action);
+}
+
+// ============================================================================
+// Shared Compensation Actions
+// ============================================================================
+
+/// Compensation action to revert proof reservation.
+///
+/// This compensation is used when a saga fails after proofs have been reserved.
+/// It sets the proofs back to Unspent state and deletes the saga record.
+///
+/// Used by: send, melt, and swap sagas.
+pub struct RevertProofReservation {
+    /// Database reference
+    pub localstore: Arc<dyn WalletDatabase<database::Error> + Send + Sync>,
+    /// Y values (public keys) of the reserved proofs
+    pub proof_ys: Vec<PublicKey>,
+    /// Saga ID for cleanup
+    pub saga_id: uuid::Uuid,
+}
+
+#[async_trait]
+impl CompensatingAction for RevertProofReservation {
+    #[instrument(skip_all)]
+    async fn execute(&self) -> Result<(), Error> {
+        tracing::info!(
+            "Compensation: Reverting {} proofs from Reserved to Unspent",
+            self.proof_ys.len()
+        );
+
+        self.localstore
+            .update_proofs_state(self.proof_ys.clone(), State::Unspent)
+            .await
+            .map_err(Error::Database)?;
+
+        // Delete saga record (best-effort)
+        if let Err(e) = self.localstore.delete_saga(&self.saga_id).await {
+            tracing::warn!(
+                "Compensation: Failed to delete saga {}: {}. Will be cleaned up on recovery.",
+                self.saga_id,
+                e
+            );
+            // Don't fail compensation if saga deletion fails - orphaned saga is harmless
+        }
+
+        Ok(())
+    }
+
+    fn name(&self) -> &'static str {
+        "RevertProofReservation"
+    }
+}
+
+/// Compensation action to revert swapped proofs.
+///
+/// This compensation is used when a saga fails after a nested swap has succeeded.
+/// The swap created new proofs that need to be released back to Unspent state.
+/// Unlike RevertProofReservation, this does NOT delete the saga (that's handled
+/// separately).
+///
+/// Used by: send and melt sagas.
+pub struct RevertSwappedProofs {
+    /// Database reference
+    pub localstore: Arc<dyn WalletDatabase<database::Error> + Send + Sync>,
+    /// Y values (public keys) of the swapped proofs
+    pub proof_ys: Vec<PublicKey>,
+}
+
+#[async_trait]
+impl CompensatingAction for RevertSwappedProofs {
+    #[instrument(skip_all)]
+    async fn execute(&self) -> Result<(), Error> {
+        tracing::info!(
+            "Compensation: Reverting {} swapped proofs to Unspent",
+            self.proof_ys.len()
+        );
+
+        self.localstore
+            .update_proofs_state(self.proof_ys.clone(), State::Unspent)
+            .await
+            .map_err(Error::Database)?;
+
+        Ok(())
+    }
+
+    fn name(&self) -> &'static str {
+        "RevertSwappedProofs"
+    }
 }
 
 #[cfg(test)]
