+++ crates/cdk/src/wallet/swap/saga/compensation.rs
@@ -3,64 +3,14 @@
 //! When a saga step fails, compensating actions are executed in reverse order (LIFO)
 //! to undo all completed steps and restore the database to its pre-saga state.
 
-use std::sync::Arc;
-
-use async_trait::async_trait;
-use cdk_common::database::{self, WalletDatabase};
-use tracing::instrument;
-
-use crate::nuts::{PublicKey, State};
-use crate::wallet::saga::CompensatingAction;
-use crate::Error;
-
-/// Compensation action to revert proof reservation for swap.
-///
-/// This compensation is used when swap fails after proofs have been reserved.
-/// It sets the proofs back to Unspent state and deletes the saga record.
-pub struct RevertSwapProofReservation {
-    /// Database reference
-    pub localstore: Arc<dyn WalletDatabase<database::Error> + Send + Sync>,
-    /// Y values (public keys) of the reserved proofs
-    pub proof_ys: Vec<PublicKey>,
-    /// Saga ID for cleanup
-    pub saga_id: uuid::Uuid,
-}
-
-#[async_trait]
-impl CompensatingAction for RevertSwapProofReservation {
-    #[instrument(skip_all)]
-    async fn execute(&self) -> Result<(), Error> {
-        tracing::info!(
-            "Compensation: Reverting {} swap proofs from Reserved to Unspent",
-            self.proof_ys.len()
-        );
-
-        self.localstore
-            .update_proofs_state(self.proof_ys.clone(), State::Unspent)
-            .await
-            .map_err(Error::Database)?;
-
-        // Delete saga record (best-effort)
-        if let Err(e) = self.localstore.delete_saga(&self.saga_id).await {
-            tracing::warn!(
-                "Compensation: Failed to delete saga {}: {}. Will be cleaned up on recovery.",
-                self.saga_id,
-                e
-            );
-            // Don't fail compensation if saga deletion fails - orphaned saga is harmless
-        }
-
-        Ok(())
-    }
-
-    fn name(&self) -> &'static str {
-        "RevertSwapProofReservation"
-    }
-}
+// Re-export shared compensation action used by swap saga
+// Note: RevertSwapProofReservation was renamed to the shared RevertProofReservation
+pub use crate::wallet::saga::RevertProofReservation as RevertSwapProofReservation;
 
 #[cfg(test)]
 mod tests {
     use std::str::FromStr;
+    use std::sync::Arc;
 
     use cdk_common::database::WalletDatabase;
     use cdk_common::nuts::{CurrencyUnit, Id, Proof, State};
@@ -71,6 +21,7 @@ mod tests {
     use cdk_common::{Amount, SecretKey};
 
     use super::*;
+    use crate::wallet::saga::CompensatingAction;
 
     /// Create test database
     async fn create_test_db() -> Arc<dyn WalletDatabase<cdk_common::database::Error> + Send + Sync>
